# If memory maximum is between bounds, make it static