--- a/lib/runtime-core/src/types.rs
+++ b/lib/runtime-core/src/types.rs
@@ -345,7 +345,7 @@ impl MemoryDescriptor {
             (true, None) => {
                 return Err("Maximum number of pages is required for shared memory".to_string())
             }
-            (false, Some(maximum)) if maximum <= STATIC_MEMORY_BOUND => MemoryType::Static,
+            (false, Some(max)) if max <= STATIC_MEMORY_BOUND => MemoryType::Static,
             (false, _) => MemoryType::Dynamic,
         })
     }