<!-- 
Prior to submitting a PR, review the CONTRIBUTING.md document for recommendations on how to test:
https://github.com/wasmerio/wasmer/blob/master/CONTRIBUTING.md#pull-requests

-->

# Description

This PR improves clif compilation speed in some cases up to 3x. Runtime speed about 40% faster.

## With the patch:
Compilation (3.376s)
```
➜  wasmer git:(master) ✗ ./target/release/wasmer cache clean && time WAPM_RUNTIME=./target/release/wasmer wax hello-go
Hello world!
WAPM_RUNTIME=./target/release/wasmer wax hello-go  8.04s user 0.18s system 243% cpu 3.376 total
```

Coremark (14549 iterations/sec)
```
➜  rusty_v8 git:(master) ✗ time ../wasmer/target/release/wasmer /var/folders/l5/vjn0dbf96fx3rd6d2vx2mcxm0000gn/T/wax/vshymanskyy/coremark@0.0.1/wapm_packages/vshymanskyy/coremark@0.0.1/build/coremark.wasm --disable-cache
2K performance run parameters for coremark.
CoreMark Size    : 666
Total ticks      : 13746
Total time (secs): 13.746000
Iterations/Sec   : 14549.687182
Iterations       : 200000
Compiler version : Clang 9.0.0 (https://github.com/llvm/llvm-project 0399d5a9682b3cef71c653373e38890c63c4c365)
Compiler flags   : -O3   -lrt
Memory location  : HEAP
seedcrc          : 0xe9f5
[0]crclist       : 0xe714
[0]crcmatrix     : 0x1fd7
[0]crcstate      : 0x8e3a
[0]crcfinal      : 0x4983
Correct operation validated. See README.md for run and reporting rules.
CoreMark 1.0 : 14549.687182 / Clang 9.0.0 (https://github.com/llvm/llvm-project 0399d5a9682b3cef71c653373e38890c63c4c365) -O3   -lrt / HEAP
../wasmer/target/release/wasmer  --disable-cache  21.34s user 0.06s system 99% cpu 21.452 total
```

## Without the patch:
Compilation (8.947s)
```
➜  wasmer git:(master) ✗ wasmer cache clean && time wax hello-go
Hello world!
wax hello-go  17.16s user 0.30s system 195% cpu 8.947 total
```

Coremark (10083 iterations/sec)
```
➜  rusty_v8 git:(master) ✗ wax coremark
2K performance run parameters for coremark.
CoreMark Size    : 666
Total ticks      : 19835
Total time (secs): 19.835000
Iterations/Sec   : 10083.186287
Iterations       : 200000
Compiler version : Clang 9.0.0 (https://github.com/llvm/llvm-project 0399d5a9682b3cef71c653373e38890c63c4c365)
Compiler flags   : -O3   -lrt
Memory location  : HEAP
seedcrc          : 0xe9f5
[0]crclist       : 0xe714
[0]crcmatrix     : 0x1fd7
[0]crcstate      : 0x8e3a
[0]crcfinal      : 0x4983
Correct operation validated. See README.md for run and reporting rules.
CoreMark 1.0 : 10083.186287 / Clang 9.0.0 (https://github.com/llvm/llvm-project 0399d5a9682b3cef71c653373e38890c63c4c365) -O3   -lrt / HEAP
```

<!-- 
Provide details regarding the change including motivation,
links to related issues, and the context of the PR.
-->

# Review

- [ ] Add a short description of the the change to the CHANGELOG.md file