@syrusakbary 2020-08-17
This issue should be fixed in wasmer `master` branch. Closing PR!